requirements.txt
slack-bolt
mysql-connector-python